Expanding on the character controller and adding more testing assets

+++ Assets/Scripts/Move.cs
@@ -3,7 +3,7 @@ public class Move
 {
     public MoveState moveState;
     public string animationName;
-    public float damage;
+    public int damage;
     public int windupTime;
     public int activeTime;
     public int basicRecovery;
@@ -11,7 +11,7 @@ public class Move
     public int hitStunTime;
     public int blockStunTime;
 
-    public Move(string anim, float dmg, int windup, int active, int recovery, int buffer, int hitStun, int blockStun)
+    public Move(string anim, int dmg, int windup, int active, int recovery, int buffer, int hitStun, int blockStun)
     {
         moveState = MoveState.Waiting;
         animationName = anim;
